services/ai-engine/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
redis>=5.0.1
pymongo>=4.5.0
numpy>=1.26.0
pandas>=2.1.2
scikit-learn>=1.3.2
transformers>=4.35.0
torch>=2.1.0
spacy>=3.7.2
python-multipart==0.0.6
google-generativeai==0.3.1
